When converting to an "Exponential" ValueNode, default the exponent to zero so the value remains the same.
git-svn-id: http://svn.voria.com/code@1883 1f10aa63-cdf2-0310-b900-c93c546f37ac

+++ synfig-core/trunk/src/synfig/valuenode_exp.cpp
@@ -56,7 +56,7 @@ ValueNode_Exp::ValueNode_Exp(const ValueBase &value):
 	switch(value.get_type())
 	{
 	case ValueBase::TYPE_REAL:
-		set_link("exp",ValueNode_Const::create(Real(1)));
+		set_link("exp",ValueNode_Const::create(Real(0)));
 		set_link("scale",ValueNode_Const::create(value.get(Real())));
 		break;
 	default:
